+++ Roguelike/Roguelike/src/Game.cs
@@ -6,13 +6,12 @@ namespace Roguelike {
         public Time Time;
         
         public Game() {
-            
             Date = new Date(31, 12, 2019);    
             Time = new Time(hours: 6, minutes: 14, seconds: 19);
             
             Cave cave = new Cave(100, 100);
             cave.Generate();
-            
+        
             while (true) {
                 Console.Clear();
                 
@@ -20,7 +19,7 @@ namespace Roguelike {
                 Console.Write("     ");
                 Time.Print(true);
 
-                ushort daysPassed = (ushort)Time.PassTime(hours: 1);
+                int daysPassed = (int)Time.PassTime(hours: 1);
                 Date.PassTime(days: daysPassed);
 
                 Console.ReadKey();
